requirements.txt
langchain
langchain-community
langchain-text-splitters
chromadb
python-dotenv
openai